Fleshing out the transport manager class with some basic functionality and examples of javadoc use.
