+++ LVT/src/systemsGo.java
@@ -1,0 +1,10 @@
+public class systemsGo
+{
+    Location br1, br2, is1, is2, ml1, ml2;
+    TransportManager tpMgr;
+
+    public static void main(String[] Args) {
+
+    }
+
+}
